+++ docs/CHANGELOG.md
@@ -17,6 +17,71 @@ All updates are (downwards) compatible! If not this is listed in the important r
 
 
 
+
+
+## [7.6.4] - 2024-06-19
+
+### Important Notes:
+ 
+
+
+### Changed
+**[TRANSLATE-4015](https://jira.translate5.net/browse/TRANSLATE-4015): Export - Error on export of SDLXLIFF with inserted tag** <br>
+Fix: SDLXLIFF track-changes export of inserted tag
+
+**[TRANSLATE-4003](https://jira.translate5.net/browse/TRANSLATE-4003): Okapi integration - Okapi server config only on system level** <br>
+The Okapi server config is changed to system level, so that it can only be changed via CLI interface and not anymore over the UI config.
+
+
+### Bugfixes
+**[TRANSLATE-4012](https://jira.translate5.net/browse/TRANSLATE-4012): Editor general - RootCause: Cannot read properties of undefined (reading 'replace')** <br>
+translate-7.6.4: Implement logging to better trace this problem.
+
+**[TRANSLATE-4004](https://jira.translate5.net/browse/TRANSLATE-4004): Installation & Update - switching http https context on using sessionTokens** <br>
+When accessing /editor instead /editor/ a redirect to http was made also in https context which might break translate5 integration scenarios 
+
+**[TRANSLATE-4001](https://jira.translate5.net/browse/TRANSLATE-4001): Workflows - ArchiveTaskActions may be stuck on old tasks and loose data** <br>
+ArchiveTaskActions does not archive tasks if there are old tasks in state error and workflowstep filter is used.
+
+**[TRANSLATE-3999](https://jira.translate5.net/browse/TRANSLATE-3999): Auto-QA - Line length evaluation in Length check fires warnings** <br>
+Fix: Line length evaluation in Length check fires warnings
+
+**[TRANSLATE-3968](https://jira.translate5.net/browse/TRANSLATE-3968): API, Authentication - Internal API not usable with application tokens** <br>
+7.6.4: Fix for sessions created via API token
+7.6.2: Functionality which is using the Internal API is not usable with application tokens.
+
+**[TRANSLATE-3924](https://jira.translate5.net/browse/TRANSLATE-3924): TermPortal - TermTranslation terms do not appear in TermCollection** <br>
+FIXED: problem with reimport translated term back to termcollection
+
+
+## [7.6.3] - 2024-06-10
+
+### Important Notes:
+ 
+
+
+### Added
+**[TRANSLATE-4000](https://jira.translate5.net/browse/TRANSLATE-4000): Editor general - Simple json editor for UI configs** <br>
+New simple json editor in the UI for map configs.
+
+
+### Changed
+**[TRANSLATE-3923](https://jira.translate5.net/browse/TRANSLATE-3923): Auto-QA - "Not found in target" category according to target term** <br>
+translate5 - 7.5.0: Quality errors in 'Not found in target' category group now count cases when best possible translations of source terms are not found in segment target
+translate5 - 7.6.3: Improve tests
+
+
+### Bugfixes
+**[TRANSLATE-3998](https://jira.translate5.net/browse/TRANSLATE-3998): Export - Wrong date values in excel export** <br>
+Fixed wrong dates in excel export when the date time is 00:00:00
+
+**[TRANSLATE-2500](https://jira.translate5.net/browse/TRANSLATE-2500): Main back-end mechanisms (Worker, Logging, etc.) - Worker Architecture: Solving Problems with Deadlocks and related Locking/Mutex Quirks** <br>
+5.2.2 Improved the internal worker handling regarding DB dead locks and a small opportunity that workers run twice.
+7.5.0 Improved the setRunning condition to reduce duplicated worker runs
+7.6.3 Improved worker queue for large project imports
+
+
+
 ## [7.6.2] - 2024-06-07
 
 ### Important Notes: